refactor(canvas): deepen chat composer input
